fix: refactoring component isolating radix
