+++ components/ModalTrigger/styles.ts
@@ -1,4 +1,3 @@
-import React from 'react';
 import { styled } from '../../styles/stitches.config';
 import { keyframes } from '@stitches/react';
 import * as AlertDialogPrimitive from '@radix-ui/react-alert-dialog';
@@ -39,20 +38,4 @@ export const StyledContent = styled(AlertDialogPrimitive.Content, {
     '&:focus': { outline: 'none' },
 });
 
-export const StyledDescription = styled(AlertDialogPrimitive.Description, {
-    margin: '20px',
-    color: '$base',
-    fontSize: '1.5rem',
-    lineHeight: '1.5',
-    textAlign: 'center'
-});
-
-export const ButtonsContainer = styled('div', { display: 'flex', padding: '0 20px' });
-
 export const TriggerButton = styled('button', { border: 'none', backgroundColor: 'transparent' })
-
-export const Content = styled('div', {
-    display: 'flex',
-    flexWrap: 'wrap',
-    justifyContent: 'center'
-})
