+++ pages/home/index.tsx
@@ -1,3 +1,11 @@
+import CategoryCard from "../../components/CategoryCard";
+import ClassCard from "../../components/ClassCard";
+import DeleteComponent from "../../components/DeleteComponent";
+
+
+
 export default function Home() {
-  return <div>Hello World</div>
+  return <div style={{backgroundColor: 'black', height: '100vh'}}>
+    <DeleteComponent/>
+  </div>
 }
